LayoutTests/http/tests/device-orientation/device-orientation-permission-granted-in-other-iframes-from-same-origin-expected.txt

CONSOLE MESSAGE: No device orientation events will be fired, reason: Permission to use the API was not yet requested.
Received device orientation & motion access request for top level origin "http://127.0.0.1:8000", with frame origin "http://localhost:8000".
Tests that giving device orientation permission to one third-party iframe gives it to other iframes of the same origin too.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.data is "granted"
PASS e.data is "granted"
PASS successfullyParsed is true

TEST COMPLETE
